backend/requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.7
PyJWT==1.7.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
django-debug-toolbar==4.2.0
drf-yasg==1.21.7

boto3==1.28.84
botocore==1.31.84
django-storages==1.14.2

inflection==0.5.1
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
python-dateutil==2.8.2
s3transfer==0.7.0
six==1.16.0
uritemplat==4.1.1
urllib3==2.0.7